requirements.txt
# Telegram Bot
aiogram==3.5.0

# OpenAI API Client
openai==1.25.1

# Environment configuration
python-dotenv==1.0.1

# HTTP clients
requests==2.32.3
aiohttp==3.9.5


